refactor config import path for sensor listener
